% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ccc_list.R
\docType{data}
\name{ccc_list}
\alias{ccc_list}
\title{California Community Colleges address data set}
\format{
A data frame with the following variables:
\describe{
  \item{College}{College name.}
  \item{District}{District name.}
  \item{Address}{Address of the college.}
  \item{Phone}{Phone number of the college.}
  \item{Website}{Web address of the college.}
}
}
\usage{
data(ccc_list)
}
\description{
List of California Community Colleges (CCC) and their addresses, obtained from \href{https://www.cccco.edu/Students/Find-a-College/College-Alphabetical-Listing}{this} site on August 31, 2021.
}
\examples{
data(ccc_list)
}
\keyword{datasets}
