% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc_list.R
\docType{data}
\name{uc_list}
\alias{uc_list}
\title{University of California address data set}
\format{
A data frame with the following variables:
\describe{
  \item{University}{University name.}
  \item{Address}{Address of the university.}
}
}
\usage{
data(uc_list)
}
\description{
List of Universities of California (UC) and their addresses, gathered on September 1, 2021.
}
\examples{
data(uc_list)
}
\keyword{datasets}
